# docs/api/order-service.openapi.yaml
openapi: 3.1.0
info:
  title: Order Service API
  version: 1.0.0
  description: |
    OrderFlow Order Service — handles authentication and order management.
    All endpoints require JWT (RS256) except /v1/auth/register and /v1/auth/login.
  contact:
    name: OrderFlow Team
  license:
    name: UNLICENSED

servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://api.orderflow.internal/order-service
    description: Production (internal ALB)

tags:
  - name: Auth
    description: Authentication and user management
  - name: Orders
    description: Order CRUD operations

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    RegisterRequest:
      type: object
      required: [email, password, consentTimestamp]
      properties:
        email:
          type: string
          format: email
          maxLength: 255
        password:
          type: string
          minLength: 8
          maxLength: 72
        consentTimestamp:
          type: string
          format: date-time
          description: GDPR consent timestamp (ISO 8601)

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    AuthResponse:
      type: object
      required: [accessToken, refreshToken, expiresIn]
      properties:
        accessToken:
          type: string
          description: JWT RS256 access token (15 min TTL)
        refreshToken:
          type: string
          description: Opaque refresh token (7 day TTL)
        expiresIn:
          type: integer
          description: Access token TTL in seconds
          example: 900

    OrderStatus:
      type: string
      enum: [pending, confirmed, shipped, delivered, cancelled]

    CreateOrderRequest:
      type: object
      required: [itemName, quantity]
      properties:
        itemName:
          type: string
          minLength: 1
          maxLength: 255
        quantity:
          type: integer
          minimum: 1
          maximum: 9999
        notes:
          type: string
          maxLength: 1000
        idempotencyKey:
          type: string
          format: uuid
          description: Client-generated idempotency key to prevent duplicate orders

    Order:
      type: object
      required: [id, userId, itemName, quantity, status, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        itemName:
          type: string
        quantity:
          type: integer
        notes:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/OrderStatus'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    OrderListResponse:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Order'
        pagination:
          type: object
          required: [hasNextPage]
          properties:
            hasNextPage:
              type: boolean
            nextCursor:
              type: string
              nullable: true
              description: Opaque cursor for next page

    UpdateOrderStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          $ref: '#/components/schemas/OrderStatus'

    AuditEntry:
      type: object
      required: [id, orderId, userId, action, fromStatus, toStatus, timestamp]
      properties:
        id:
          type: string
          format: uuid
        orderId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        action:
          type: string
        fromStatus:
          $ref: '#/components/schemas/OrderStatus'
        toStatus:
          $ref: '#/components/schemas/OrderStatus'
        timestamp:
          type: string
          format: date-time

    HealthResponse:
      type: object
      required: [status, timestamp]
      properties:
        status:
          type: string
          enum: [ok, degraded, down]
        timestamp:
          type: string
          format: date-time
        version:
          type: string

    ErrorResponse:
      type: object
      required: [error, message, correlationId]
      properties:
        error:
          type: string
        message:
          type: string
        correlationId:
          type: string
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string

  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnprocessableEntity:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

security:
  - bearerAuth: []

paths:
  /health:
    get:
      summary: Liveness probe
      tags: [Auth]
      security: []
      operationId: getHealth
      responses:
        '200':
          description: Service is alive
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /ready:
    get:
      summary: Readiness probe
      tags: [Auth]
      security: []
      operationId: getReadiness
      responses:
        '200':
          description: Service is ready to accept traffic
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
        '503':
          description: Service not ready (DB or deps unavailable)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /v1/auth/register:
    post:
      summary: Register a new user
      tags: [Auth]
      security: []
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'

  /v1/auth/login:
    post:
      summary: Login and receive JWT tokens
      tags: [Auth]
      security: []
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'

  /v1/auth/me:
    delete:
      summary: GDPR right-to-deletion — delete own account
      tags: [Auth]
      operationId: deleteAccount
      responses:
        '204':
          description: Account deleted
        '401':
          $ref: '#/components/responses/Unauthorized'

  /v1/orders:
    post:
      summary: Create a new order
      tags: [Orders]
      operationId: createOrder
      parameters:
        - in: header
          name: Idempotency-Key
          schema:
            type: string
            format: uuid
          description: Idempotency key to prevent duplicate orders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'

    get:
      summary: List user's orders (cursor pagination)
      tags: [Orders]
      operationId: listOrders
      parameters:
        - in: query
          name: cursor
          schema:
            type: string
          description: Opaque cursor from previous response
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/OrderStatus'
      responses:
        '200':
          description: Order list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /v1/orders/{id}:
    get:
      summary: Get order detail
      tags: [Orders]
      operationId: getOrder
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Order detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /v1/orders/{id}/status:
    patch:
      summary: Update order status
      tags: [Orders]
      operationId: updateOrderStatus
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderStatusRequest'
      responses:
        '200':
          description: Status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Invalid status transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'